Add files via upload
An example dictionary that loads a data file (data.json) and provides a simple search in the database for words based on user input. Dictionary outputs the definitions of said words.

The dictionary has been designed to search for close matches to words as well as mistakes with upper and lower case.